requirements.txt
numpy~=1.26.4
matplotlib~=3.8.3
torch~=2.2.0+cu118
torchvision~=0.17.0+cu118
pyaer~=0.2.7a0
qpsolvers~=4.3.1
scipy~=1.11.4
scikit-learn~=1.4.1.post1
pyserial~=3.5
pillow~=10.2.0